Mostly finished the linkstate packet, john need
John, create the actual connections to the things
also, where are we updating the timestamps from?
I modified the connection_node and the node structs a bit, to make their
fields proper, and to add the socket to the node struct.
Other than that, I think that it works, I'll see if I can write some
more of the things that we need to change